+++ Assets/Scripts/Models/Events, Repos/ScoreUserPrefs.cs
@@ -3,7 +3,7 @@
     public string Username;
     public int Position;
     public int Score;
-    public float Timer;
+    public int Timer;
 
     public ScoreUserPrefs()
     {
@@ -13,18 +13,18 @@
     {
         Username = userName;
     }
-    public ScoreUserPrefs(int score, float timer)
+    public ScoreUserPrefs(int score, int timer)
     {
         Score = score;
         Timer = timer;
     }
-    public ScoreUserPrefs(string username, int score, float timer)
+    public ScoreUserPrefs(string username, int score, int timer)
     {
         Username = username;
         Score = score;
         Timer = timer;
     }
-    public ScoreUserPrefs(string username, int position, int score, float timer)
+    public ScoreUserPrefs(string username, int position, int score, int timer)
     {
         Username = username;
         Position = position;
